refactor: reorganize genai module structure and enhance documentation
